BREAKING(collections): move RedBlackTree, BinarySearchTree, and BinaryHeap to 'unstable' subdir
