+++ collections/unstable/binary_heap_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assertEquals } from "../assert/mod.ts";
-import { assert } from "../assert/assert.ts";
-import { ascend, BinaryHeap, descend } from "./binary_heap.ts";
+import { assert, assertEquals } from "../../assert/mod.ts";
+import { BinaryHeap } from "./binary_heap.ts";
+import { ascend, descend } from "./comparators.ts";
 import { Container, MyMath } from "./_test_utils.ts";
 
 Deno.test("[collections/BinaryHeap] with default descend comparator", () => {
