+++ collections/unstable/red_black_tree_test.ts
@@ -1,6 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assertEquals, assertStrictEquals } from "../assert/mod.ts";
-import { ascend, descend, RedBlackTree } from "./red_black_tree.ts";
+import { assertEquals, assertStrictEquals } from "../../assert/mod.ts";
+import { RedBlackTree } from "./red_black_tree.ts";
+import { ascend, descend } from "./comparators.ts";
 import { Container, MyMath } from "./_test_utils.ts";
 
 Deno.test("[collections/RedBlackTree] with default ascend comparator", () => {
